fix: clippy gate rename in ulua-require; soundness fixes in code-gen/compiler/bytecode
- ulua-require: rename Navigator to_alias_* -> navigate_to_alias_* (wrong_self_convention)
- ulua-code-gen: emit_add_offset bounds-check before u16 truncation
- ulua-compiler: userdata count limit check in usize instead of wrapping u16
- ulua-bytecode: BytecodeBuilder::default() delegates to new(None) so constant_map sentinel no longer collides with nil keys

+++ crates/ulua-code-gen/src/functions/emit_add_offset.rs
@@ -14,7 +14,7 @@ pub fn emit_add_offset(
 
   const K_MAX_IMMEDIATE: u16 = 4095;
 
-  if offset as u16 <= K_MAX_IMMEDIATE {
+  if offset <= K_MAX_IMMEDIATE as usize {
     build.add_register_a_64_register_a_64_u16(dst, src, offset as u16);
   } else {
     build.mov_register_a_64_i32(dst, offset as i32);
